+++ src/@types/vscode-variables/index.d.ts
@@ -1,0 +1,4 @@
+declare module 'vscode-variables' {
+  function vscodeVariables(str: string, recursive?: boolean): string;
+  export = vscodeVariables;
+}